add another rtree as index for the sensor position

+++ seerep_msgs/core/polygon2d.h
@@ -8,8 +8,8 @@ namespace seerep_core_msgs
 struct Polygon2D
 {
   std::vector<Point2D> vertices;
-  int z;
-  int height;
+  double z;
+  double height;
 };
 }  // namespace seerep_core_msgs
